requirements.txt
Flask
Flask-SQLAlchemy
Flask-Bcrypt
Flask-Login
pyotp
qrcode
Pillow
python-dotenv
cryptography